Lineage Editor, pt 2
#783
