+++ waltz-ng/client/applications/app-view.html
@@ -209,8 +209,7 @@
                      select="ctrl.dataTab = 1"
                      heading="Physical Flow Detail">
                 <waltz-physical-data-section specifications="ctrl.physicalSpecifications"
-                                             physical-flows="ctrl.physicalFlows"
-                                             endpoint-references="ctrl.physicalEndpointReferences">
+                                             physical-flows="ctrl.physicalFlows">
                 </waltz-physical-data-section>
             </uib-tab>
         </uib-tabset>
