+++ waltz-ng/client/physical-specifications/components/physical-data-section/physical-data-section.js
@@ -1,10 +1,8 @@
-import _ from 'lodash';
-import {combineFlowData} from '../../utilities';
+import {combineFlowData, enrichConsumes} from '../../utilities';
 import {termSearch} from "../../../common"
 
 
 const bindings = {
-    endpointReferences: '<',  // [ entityReference... ]
     physicalFlows: '<',
     specifications: '<'
 };
@@ -13,55 +11,15 @@ const bindings = {
 const template = require('./physical-data-section.html');
 
 
-function enrichConsumes(specifications = [],
-                physicalFlows = [],
-                endpointReferences = [])
-{
-    const visitedRefs = [];
-
-    return _.chain(specifications)
-        .map(specification => {
-            const physicalFlow = _.find(physicalFlows, { specificationId: specification.id });
-            const sourceRef = _.find(endpointReferences, { id: specification.owningEntity.id, kind: specification.owningEntity.kind });
-            const targetRef = _.find(endpointReferences, { id: physicalFlow.target.id, kind: physicalFlow.target.kind });
-
-            if (!physicalFlow || !sourceRef || !targetRef) {
-                return null;
-            } else {
-                const firstSource = !_.includes(
-                    visitedRefs,
-                    sourceRef);
-
-                if(firstSource === true) {
-                    visitedRefs.push(sourceRef);
-                }
-
-                return {
-                    specification,
-                    physicalFlow,
-                    firstSource,
-                    sourceRef,
-                    targetRef
-                };
-            }
-        })
-        .filter(r => r != null)
-        .value();
-}
-
-
 function mkData(specifications = { produces: [], consumes: [] },
-                physicalFlows = [],
-                endpointReferences = [])
+                physicalFlows = [])
 {
     const produces = combineFlowData(
         specifications.produces,
-        physicalFlows,
-        endpointReferences);
+        physicalFlows);
     const consumes = enrichConsumes(
         specifications.consumes,
-        physicalFlows,
-        endpointReferences);
+        physicalFlows);
     return { produces, consumes };
 }
 
@@ -92,7 +50,7 @@ function controller() {
     ];
 
     vm.$onChanges = (changes) => {
-        Object.assign(vm, mkData(vm.specifications, vm.physicalFlows, vm.endpointReferences));
+        Object.assign(vm, mkData(vm.specifications, vm.physicalFlows));
         vm.filterProduces("");
         vm.filterConsumes("");
     };
